+++ internal/db/org_members.go
@@ -5,56 +5,96 @@ import (
 	"database/sql"
 	"errors"
 	"fmt"
+	"sync"
 
+	"github.com/jackc/pgconn"
 	"github.com/keegancsmith/sqlf"
-	"github.com/lib/pq"
 
 	"github.com/sourcegraph/sourcegraph/internal/db/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/db/dbconn"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbutil"
 	"github.com/sourcegraph/sourcegraph/internal/types"
 )
 
-type orgMembers struct{}
+type OrgMemberStore struct {
+	*basestore.Store
 
-func (*orgMembers) Create(ctx context.Context, orgID, userID int32) (*types.OrgMembership, error) {
-	m := types.OrgMembership{
+	once sync.Once
+}
+
+// NewOrgMemberStoreWithDB instantiates and returns a new OrgMemberStore with prepared statements.
+func NewOrgMemberStoreWithDB(db dbutil.DB) *OrgMemberStore {
+	return &OrgMemberStore{Store: basestore.NewWithDB(db, sql.TxOptions{})}
+}
+
+// NewOrgMemberStoreWithDB instantiates and returns a new OrgMemberStore using the other store handle.
+func NewOrgMemberStoreWith(other basestore.ShareableStore) *OrgMemberStore {
+	return &OrgMemberStore{Store: basestore.NewWithHandle(other.Handle())}
+}
+
+func (s *OrgMemberStore) With(other basestore.ShareableStore) *OrgMemberStore {
+	return &OrgMemberStore{Store: s.Store.With(other)}
+}
+
+func (m *OrgMemberStore) Transact(ctx context.Context) (*OrgMemberStore, error) {
+	txBase, err := m.Store.Transact(ctx)
+	return &OrgMemberStore{Store: txBase}, err
+}
+
+// ensureStore instantiates a basestore.Store if necessary, using the dbconn.Global handle.
+// This function ensures access to dbconn happens after the rest of the code or tests have
+// initialized it.
+func (m *OrgMemberStore) ensureStore() {
+	m.once.Do(func() {
+		if m.Store == nil {
+			m.Store = basestore.NewWithDB(dbconn.Global, sql.TxOptions{})
+		}
+	})
+}
+
+func (m *OrgMemberStore) Create(ctx context.Context, orgID, userID int32) (*types.OrgMembership, error) {
+	m.ensureStore()
+
+	om := types.OrgMembership{
 		OrgID:  orgID,
 		UserID: userID,
 	}
-	err := dbconn.Global.QueryRowContext(
+	err := m.Handle().DB().QueryRowContext(
 		ctx,
 		"INSERT INTO org_members(org_id, user_id) VALUES($1, $2) RETURNING id, created_at, updated_at",
-		m.OrgID, m.UserID).Scan(&m.ID, &m.CreatedAt, &m.UpdatedAt)
+		om.OrgID, om.UserID).Scan(&om.ID, &om.CreatedAt, &om.UpdatedAt)
 	if err != nil {
-		if pqErr, ok := err.(*pq.Error); ok {
-			if pqErr.Constraint == "org_members_org_id_user_id_key" {
+		if pgErr, ok := err.(*pgconn.PgError); ok {
+			if pgErr.ConstraintName == "org_members_org_id_user_id_key" {
 				return nil, errors.New("user is already a member of the organization")
 			}
 		}
 		return nil, err
 	}
-	return &m, nil
+	return &om, nil
 }
 
-func (m *orgMembers) GetByUserID(ctx context.Context, userID int32) ([]*types.OrgMembership, error) {
+func (m *OrgMemberStore) GetByUserID(ctx context.Context, userID int32) ([]*types.OrgMembership, error) {
 	return m.getBySQL(ctx, "INNER JOIN users ON org_members.user_id=users.id WHERE org_members.user_id=$1 AND users.deleted_at IS NULL", userID)
 }
 
-func (m *orgMembers) GetByOrgIDAndUserID(ctx context.Context, orgID, userID int32) (*types.OrgMembership, error) {
+func (m *OrgMemberStore) GetByOrgIDAndUserID(ctx context.Context, orgID, userID int32) (*types.OrgMembership, error) {
 	if Mocks.OrgMembers.GetByOrgIDAndUserID != nil {
 		return Mocks.OrgMembers.GetByOrgIDAndUserID(ctx, orgID, userID)
 	}
 	return m.getOneBySQL(ctx, "INNER JOIN users ON org_members.user_id=users.id WHERE org_id=$1 AND user_id=$2 AND users.deleted_at IS NULL LIMIT 1", orgID, userID)
 }
 
-func (*orgMembers) Remove(ctx context.Context, orgID, userID int32) error {
-	_, err := dbconn.Global.ExecContext(ctx, "DELETE FROM org_members WHERE (org_id=$1 AND user_id=$2)", orgID, userID)
+func (m *OrgMemberStore) Remove(ctx context.Context, orgID, userID int32) error {
+	m.ensureStore()
+
+	_, err := m.Handle().DB().ExecContext(ctx, "DELETE FROM org_members WHERE (org_id=$1 AND user_id=$2)", orgID, userID)
 	return err
 }
 
 // GetByOrgID returns a list of all members of a given organization.
-func (*orgMembers) GetByOrgID(ctx context.Context, orgID int32) ([]*types.OrgMembership, error) {
-	org, err := Orgs.GetByID(ctx, orgID)
+func (m *OrgMemberStore) GetByOrgID(ctx context.Context, orgID int32) ([]*types.OrgMembership, error) {
+	org, err := Orgs.With(m).GetByID(ctx, orgID)
 	if err != nil {
 		return nil, err
 	}
@@ -73,7 +113,7 @@ func (err *ErrOrgMemberNotFound) Error() string {
 
 func (ErrOrgMemberNotFound) NotFound() bool { return true }
 
-func (m *orgMembers) getOneBySQL(ctx context.Context, query string, args ...interface{}) (*types.OrgMembership, error) {
+func (m *OrgMemberStore) getOneBySQL(ctx context.Context, query string, args ...interface{}) (*types.OrgMembership, error) {
 	members, err := m.getBySQL(ctx, query, args...)
 	if err != nil {
 		return nil, err
@@ -84,7 +124,9 @@ func (m *orgMembers) getOneBySQL(ctx context.Context, query string, args ...inte
 	return members[0], nil
 }
 
-func (*orgMembers) getBySQL(ctx context.Context, query string, args ...interface{}) ([]*types.OrgMembership, error) {
+func (m *OrgMemberStore) getBySQL(ctx context.Context, query string, args ...interface{}) ([]*types.OrgMembership, error) {
+	m.ensureStore()
+
 	rows, err := dbconn.Global.QueryContext(ctx, "SELECT org_members.id, org_members.org_id, org_members.user_id, org_members.created_at, org_members.updated_at FROM org_members "+query, args...)
 	if err != nil {
 		return nil, err
@@ -108,15 +150,12 @@ func (*orgMembers) getBySQL(ctx context.Context, query string, args ...interface
 
 // CreateMembershipInOrgsForAllUsers causes *ALL* users to become members of every org in the
 // orgNames list.
-//
-// The provided dbh is used as the DB handle to execute the query. It may be either a global
-// DB handle or a transaction. If nil, the global DB handle is used.
-func (*orgMembers) CreateMembershipInOrgsForAllUsers(ctx context.Context, dbh interface {
-	Exec(ctx context.Context, query *sqlf.Query) error
-}, orgNames []string) error {
+func (m *OrgMemberStore) CreateMembershipInOrgsForAllUsers(ctx context.Context, orgNames []string) error {
 	if len(orgNames) == 0 {
 		return nil
 	}
+
+	m.ensureStore()
 
 	orgNameVars := []*sqlf.Query{}
 	for _, orgName := range orgNames {
@@ -134,9 +173,6 @@ func (*orgMembers) CreateMembershipInOrgsForAllUsers(ctx context.Context, dbh in
 			INSERT INTO org_members(org_id,user_id) SELECT to_join.org_id, to_join.user_id FROM to_join;`,
 		sqlf.Join(orgNameVars, ","))
 
-	if dbh == nil {
-		dbh = basestore.NewWithDB(dbconn.Global, sql.TxOptions{})
-	}
-	err := dbh.Exec(ctx, sqlQuery)
+	err := m.Exec(ctx, sqlQuery)
 	return err
 }