+++ internal/db/users.go
@@ -11,6 +11,7 @@ import (
 	"unicode/utf8"
 
 	"github.com/inconshreveable/log15"
+	"github.com/jackc/pgconn"
 	"github.com/keegancsmith/sqlf"
 	"github.com/lib/pq"
 	"github.com/pkg/errors"
@@ -187,7 +188,7 @@ func (u *UserStore) Create(ctx context.Context, info NewUser) (newUser *types.Us
 	if err != nil {
 		return nil, err
 	}
-	defer tx.Done(err)
+	defer func() { err = tx.Done(err) }()
 	return tx.create(ctx, info)
 }
 
@@ -279,12 +280,12 @@ func (u *UserStore) create(ctx context.Context, info NewUser) (newUser *types.Us
 		sqlf.Sprintf("INSERT INTO users(username, display_name, avatar_url, created_at, updated_at, passwd, invalidated_sessions_at, site_admin) VALUES(%s, %s, %s, %s, %s, %s, %s, %s AND NOT EXISTS(SELECT * FROM users)) RETURNING id, site_admin",
 			info.Username, info.DisplayName, avatarURL, createdAt, updatedAt, passwd, invalidatedSessionsAt, !alreadyInitialized)).Scan(&id, &siteAdmin)
 	if err != nil {
-		if pqErr, ok := err.(*pq.Error); ok {
-			switch pqErr.Constraint {
+		if pgErr, ok := err.(*pgconn.PgError); ok {
+			switch pgErr.ConstraintName {
 			case "users_username":
 				return nil, errCannotCreateUser{errorCodeUsernameExists}
 			case "users_username_max_length", "users_username_valid_chars", "users_display_name_max_length":
-				return nil, errCannotCreateUser{pqErr.Constraint}
+				return nil, errCannotCreateUser{pgErr.ConstraintName}
 			}
 		}
 		return nil, err
@@ -308,8 +309,8 @@ func (u *UserStore) create(ctx context.Context, info NewUser) (newUser *types.Us
 			err = u.Exec(ctx, sqlf.Sprintf("INSERT INTO user_emails(user_id, email, verification_code, is_primary) VALUES (%s, %s, %s, true)", id, info.Email, info.EmailVerificationCode))
 		}
 		if err != nil {
-			if pqErr, ok := err.(*pq.Error); ok {
-				switch pqErr.Constraint {
+			if pgErr, ok := err.(*pgconn.PgError); ok {
+				switch pgErr.ConstraintName {
 				case "user_emails_unique_verified_email":
 					return nil, errCannotCreateUser{errorCodeEmailExists}
 				}
@@ -340,7 +341,7 @@ func (u *UserStore) create(ctx context.Context, info NewUser) (newUser *types.Us
 		for _, err := range errs {
 			log15.Warn(err.Error())
 		}
-		if err := OrgMembers.CreateMembershipInOrgsForAllUsers(ctx, u, orgs); err != nil {
+		if err := OrgMembers.With(u).CreateMembershipInOrgsForAllUsers(ctx, orgs); err != nil {
 			return nil, err
 		}
 
@@ -393,7 +394,7 @@ func (u *UserStore) Update(ctx context.Context, id int32, update UserUpdate) (er
 	if err != nil {
 		return err
 	}
-	defer tx.Done(err)
+	defer func() { err = tx.Done(err) }()
 
 	fieldUpdates := []*sqlf.Query{
 		sqlf.Sprintf("updated_at=now()"), // always update updated_at timestamp
@@ -421,7 +422,7 @@ func (u *UserStore) Update(ctx context.Context, id int32, update UserUpdate) (er
 	query := sqlf.Sprintf("UPDATE users SET %s WHERE id=%d", sqlf.Join(fieldUpdates, ", "), id)
 	res, err := tx.ExecResult(ctx, query)
 	if err != nil {
-		if pqErr, ok := err.(*pq.Error); ok && pqErr.Constraint == "users_username" {
+		if pgErr, ok := err.(*pgconn.PgError); ok && pgErr.ConstraintName == "users_username" {
 			return errCannotCreateUser{errorCodeUsernameExists}
 		}
 		return err
@@ -447,7 +448,7 @@ func (u *UserStore) Delete(ctx context.Context, id int32) (err error) {
 	if err != nil {
 		return err
 	}
-	defer tx.Done(err)
+	defer func() { err = tx.Done(err) }()
 
 	res, err := tx.ExecResult(ctx, sqlf.Sprintf("UPDATE users SET deleted_at=now() WHERE id=%s AND deleted_at IS NULL", id))
 	if err != nil {
@@ -496,7 +497,7 @@ func (u *UserStore) HardDelete(ctx context.Context, id int32) (err error) {
 	if err != nil {
 		return err
 	}
-	defer tx.Done(err)
+	defer func() { err = tx.Done(err) }()
 
 	if err := tx.Exec(ctx, sqlf.Sprintf("DELETE FROM names WHERE user_id=%s", id)); err != nil {
 		return err
@@ -668,7 +669,7 @@ func (u *UserStore) InvalidateSessionsByID(ctx context.Context, id int32) (err e
 	if err != nil {
 		return err
 	}
-	defer tx.Done(err)
+	defer func() { err = tx.Done(err) }()
 
 	query := sqlf.Sprintf(`
 		UPDATE users